Allow for browser back/forward buttons to work
This commit introduces a "/next" endpoint which the "Next button" has
been updated to use. When the browser hits the /next endpoint it is
redirected to "?url=$url" with an HTTP 307 response code.

This allows the posts a user has visited to be stored in the browser
history making the back and forward buttons work.

I implemented this because I often accidentily skip past a piece of
content that I am interested in and am unable to find it again without
spamming the next button until I happen upon the content.